fixed nbayes training (cov matrix type)
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@857 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/ml/mlnbayes.cpp
@@ -190,7 +190,7 @@ bool CvNormalBayesClassifier::train( const CvMat* _train_data, const CvMat* _res
     }
 
     responses_data = responses->data.i;
-    CV_CALL( cov = cvCreateMat( _var_count, _var_count, CV_32FC1 ));
+    CV_CALL( cov = cvCreateMat( _var_count, _var_count, CV_64FC1 ));
 
     /* process train data (count, sum , productsum) */
     for( s = 0; s < nsamples; s++ )
